Removing old utils/index file and moving StatusNotification component back into Forms, refactored html selction using DocumentSelection component, and refactored status messages for upload/fetch/delete with runNotification
