+++ src/utils/form-helper.js
@@ -1,0 +1,1 @@
+export const docTypes = ["Bank Statement", "Passport", "Drivers License"];
